% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_metrics.R
\name{adjusted_r_squared}
\alias{adjusted_r_squared}
\title{Adjusted R Squared}
\usage{
adjusted_r_squared(known, predicted, p)
}
\arguments{
\item{known}{known values}

\item{predicted}{predicted values (point estimates)}

\item{p}{number of predictors used in regression model}
}
\value{
Adjusted R Squared value
}
\description{
Evaluates Adjusted R Squared, Explained Variance, on predictions of a
regression model.
}
\note{
A negative R Squared value means that the prediction model is worst
than fitting an horizontal line to the data.
}
