% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_descriptor.R
\name{compute_descriptor}
\alias{compute_descriptor}
\title{Compute Descriptor}
\usage{
compute_descriptor(x, fun, by = "col", ...)
}
\arguments{
\item{x}{A matrix or data.frame}

\item{fun}{A function defining a descriptor to apply to x.

fun must must operate on a vector basis.
The rows or columns of x will be passed as the first argument.
Other arguments can be passed with \strong{...}}

\item{by}{A character vector. "row" defines row-wise and "col" defines
column-wise operation.}

\item{...}{Additional arguments to pass to \emph{fun}}
}
\value{
A vector, data.frame/matrix or a list with the computed descriptor
values
}
\description{
\emph{compute_descriptor} is wrapper for the \code{\link[base]{apply}}
function to facilitate row or column wise operations on matrix and data.frame
objects.
}
\examples{
df <- iris[,-5]
# Compute the mean of iris sepal and petal dimensions. Output is a vector.
compute_descriptor(df, fun = mean, by = "col")
# Compute the range of iris sepal and petal dimensions. Output is a list
compute_descriptor(df, fun = range, by = "col")

}
