% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_grnnet.R
\name{evaluate_grnnet}
\alias{evaluate_grnnet}
\title{Evaluate \emph{grnnet} model}
\usage{
evaluate_grnnet(object, digits = 2)
}
\arguments{
\item{object}{a \emph{grnnet} object}

\item{digits}{number of digits in computed metrics}
}
\value{
relevant metrics for \emph{grnnet} output
}
\description{
Evaluate \emph{grnnet} model
}
\examples{
model <- grnnet(iris[,1:2], iris[,3:4])
evaluate_grnnet(model)

}
\author{
David Navega
}
