% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolation_interval.R
\name{extrapolation_interval}
\alias{extrapolation_interval}
\title{Extrapolation Interval}
\usage{
extrapolation_interval(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
the extrapolation interval of x
}
\description{
Extrapolation Interval
}
