% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grnnet.R
\name{grnnet}
\alias{grnnet}
\title{Generalized Regression Neural Network}
\usage{
grnnet(x, y, alpha = 0.1, crossvalidation = T, kfold = 5,
  seed = 1984)
}
\arguments{
\item{x}{a matrix of numeric inputs}

\item{y}{a matrix of numeric outputs}

\item{alpha}{a numeric value. confidence level (i.e, 0.05)}

\item{crossvalidation}{a logical indicating if k-fold cross-validation should
be performed. Default is TRUE (T)}

\item{kfold}{number of fold for k-fold cross-validation}

\item{seed}{seed for the random number generator. Controls reproducibility of
k-fold cross-validation}
}
\value{
a trained generalized regression neural network
}
\description{
Implementation of the Generalized Regression Neural Network architecture for
Single and Multi-Output Regression with Uncertainty Estimation Using
Conformal Prediction.
}
\note{
The network training is performed with an internal leave-one-out sampling
loop and derivate free optimization (Brent's algorithm). The bounds of the
sigma parameter, the spread of the squared exponential kernel, are initialized
following Tomandl & Schober (2001) based on the proprieties of the distance
matrix between cases. The internal loop also computes the conformity factors
needed to compute the prediction intervals based on conformal prediction.
}
\author{
David Navega
}
