% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kde_quantile.R
\name{kde_quantile}
\alias{kde_quantile}
\title{Quantile Estimation using Kernel Density Estimation}
\usage{
kde_quantile(x, probs = c(0.001, 0.999))
}
\arguments{
\item{x}{a numeric vector}

\item{probs}{quantiles (cumulative probability value) to assess}
}
\value{
a named vector of estimated quantiles
}
\description{
Quantile Estimation using Kernel Density Estimation
}
