% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfold_crossvalidation.R
\name{kfold_crossvalidation}
\alias{kfold_crossvalidation}
\title{K-Fold Cross-Validation}
\usage{
kfold_crossvalidation(n, k = 5, seed = NULL)
}
\arguments{
\item{n}{number of training examples}

\item{k}{number of folds. Default is 5}

\item{seed}{seed for the pseudo-random number generator}
}
\value{
a matrix of logical vectors defining observations in each fold.
}
\description{
Helper function to perform K-Fold Cross-Validation
}
