% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_metrics.R
\name{mean_absolute_error}
\alias{mean_absolute_error}
\title{Mean Absolute Error (MAE)}
\usage{
mean_absolute_error(known, predicted)
}
\arguments{
\item{known}{known values}

\item{predicted}{predicted values (point estimates)}
}
\value{
MAE value
}
\description{
Evaluates MAE on predictions of a regression model
}
