% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize Data}
\usage{
normalize(x, y = NULL)
}
\arguments{
\item{x}{a matrix or data.frame to normalize}

\item{y}{an optional matrix or data.frame previously normalized from which
normalization paramter can be extracted.}
}
\value{
matrix or data.frame with normalized data
}
\description{
Normalize data using min-max scaling. Normalized data is bounded between
0 and 1
}
