% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grnnet.R
\name{predict.grnnet}
\alias{predict.grnnet}
\title{Predict from a Generalized Regression Neural Network Object}
\usage{
\method{predict}{grnnet}(object, newdata, alpha = NULL, ...)
}
\arguments{
\item{object}{object of class 'grnnet'}

\item{newdata}{a matrix or a data.frame of new data to make predictions}

\item{alpha}{a value between 0 and 1 defining the confidence level for
conformal prediction intervals. Default is NULL and alpha is equal to
pre-defined value during training.}

\item{...}{...}
}
\value{
a matrix or a list with a point estimate and associated prediction
interval for each row of newdata. A named list is outputed if the neural
network has multiple output
}
\description{
Predict from a Generalized Regression Neural Network Object
}
\author{
David Navega
}
