% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_metrics.R
\name{prediction_bias}
\alias{prediction_bias}
\title{Regression Prediction Bias}
\usage{
prediction_bias(known, predicted)
}
\arguments{
\item{known}{known values}

\item{predicted}{predicted values (point estimates)}
}
\value{
Regression prediction bias value
}
\description{
Evaluate Regression Prediction Bias by the slope of the regression model of
residuals on known values.
}
\note{
A positive bias means that lower known values are overestimated and
upper known values are underestimated.
}
