% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_metrics.R
\name{prediction_coverage}
\alias{prediction_coverage}
\title{Prediction Coverage}
\usage{
prediction_coverage(known, predicted)
}
\arguments{
\item{known}{known values}

\item{predicted}{predicted values (prediction interval, lower and upper value)}
}
\value{
Coverage probability value
}
\description{
Evaluates the coverage of prediction intervals.
}
