% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_metrics.R
\name{prediction_interval_width}
\alias{prediction_interval_width}
\title{Prediction Interval Mean Width}
\usage{
prediction_interval_width(predicted)
}
\arguments{
\item{predicted}{predicted values (prediction interval, lower and upper value)}
}
\value{
a vector width mean width and percentiles of prediciton interval width
(0.025, 0.975)
}
\description{
Evaluates mean width of prediction intervals as a measure of efficiency
}
