% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grnnet.R
\name{print.grnnet}
\alias{print.grnnet}
\title{Print method for Generalized Regression Neural Network}
\usage{
\method{print}{grnnet}(x, digits = 2, ...)
}
\arguments{
\item{x}{a \emph{grnnet} object}

\item{digits}{digitis to print}

\item{...}{...}
}
\description{
Print method for Generalized Regression Neural Network
}
\author{
David Navega
}
