% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_euclidean_distance.R
\name{pw_euclidean_distance}
\alias{pw_euclidean_distance}
\title{Pairwise Squared Euclidean Distance}
\usage{
pw_euclidean_distance(x, y, weights = NULL)
}
\arguments{
\item{x}{a numeric matrix}

\item{y}{a numeric matrix with the same number of columns as matrix x}

\item{weights}{a numeric vector matching the number of columns defining the
weights for a weighted distance vector.}
}
\value{
a matrix with euclidean distances between rows of x and y
}
\description{
Fast Pairwise Squared Euclidean Distance Computation Using Formula Expansion.
}
\author{
David Navega
}
