% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_metrics.R
\name{r_squared}
\alias{r_squared}
\title{R Squared}
\usage{
r_squared(known, predicted)
}
\arguments{
\item{known}{known values}

\item{predicted}{predicted values (point estimates)}
}
\value{
R Squared value
}
\description{
Evaluates R Squared, Explained Variance, on predictions of a regression model
}
\note{
A negative R Squared value means that the prediction model is worst
than fitting an horizontal line to the data.
}
