% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_prediction_analysis.R
\name{regression_prediction_analysis}
\alias{regression_prediction_analysis}
\title{Prediction Analysis for Regression}
\usage{
regression_prediction_analysis(known, predicted, digits = 2,
  base_size = 14)
}
\arguments{
\item{known}{a vector of known values}

\item{predicted}{a matrix of predicted values. Matrix should be organized in
way that the first column represents the point estimate and the second and
third columns represent the lower and upper bound of the predictive interval}

\item{digits}{number of digits controling floats precision}

\item{base_size}{ggplot parameter to control text size in plots}
}
\value{
a list with the follwing compoments:
\describe{
\item{metrics}{a data.frame with the most relevant performance metrics for regression}
\item{plots}{a list storing three regression analysis plots}
}
}
\description{
Prediction Analysis for Regression
}
