% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_metrics.R
\name{root_mean_squared_error}
\alias{root_mean_squared_error}
\title{Root Mean Squared Error (RMSE)}
\usage{
root_mean_squared_error(known, predicted, nrmse = F)
}
\arguments{
\item{known}{known values}

\item{predicted}{predicted values (point estimates)}

\item{nrmse}{a logical indicating if IQR normalized rmse should be computed.
Default is FALSE (F)}
}
\value{
RMSE (or NRMSE) value
}
\description{
Evaluates RMSE on predictions of a regression model
}
