% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/squared_exponential_kernel.R
\name{squared_exponential_kernel}
\alias{squared_exponential_kernel}
\title{Squared Exponential Kernel}
\usage{
squared_exponential_kernel(x, sigma)
}
\arguments{
\item{x}{a vector or a matrix of pre-computed euclidean distances}

\item{sigma}{a numeric defining the spread of the gaussian kernel}
}
\value{
evaluation values of the squared exponential kernel function
}
\description{
Squared Exponential Kernel
}
\author{
David Navega
}
