% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_mean.R
\name{weighted_mean}
\alias{weighted_mean}
\title{Weighted Mean}
\usage{
weighted_mean(x, weights, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector}

\item{weights}{a numeric vector of weights}

\item{na.rm}{a logical for handle NA values}
}
\value{
weighted mean
}
\description{
Weighted Mean
}
