% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_metrics.R
\name{weighted_r_squared}
\alias{weighted_r_squared}
\title{Weighted R Squared}
\usage{
weighted_r_squared(known, predicted, weights)
}
\arguments{
\item{known}{known values}

\item{predicted}{predicted values (point estimates)}

\item{weights}{a vector of weights}
}
\value{
Weighted R Squared value
}
\description{
Evaluates Weighted R Squared, Explained Variance, on predictions of a
regression model.
}
\note{
A negative R Squared value means that the prediction model is worst
than fitting an horizontal line to the data.
}
