% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_sd.R
\name{weighted_sd}
\alias{weighted_sd}
\title{Weighted Standard Deviation}
\usage{
weighted_sd(x, weights, na.rm = F, w_mean)
}
\arguments{
\item{x}{a numeric vector}

\item{weights}{a numeric vector of weights}

\item{na.rm}{a logical for handle NA values}

\item{w_mean}{pre-computed mean}
}
\value{
weighted standard deviation
}
\description{
Weighted Standard Deviation
}
